pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "atoplace"
version = "0.1.0"
description = "AI-Powered PCB Placement and Routing Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "AtoPlace Team"}
]
keywords = ["pcb", "kicad", "placement", "routing", "electronics", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]

dependencies = [
    "numpy>=1.21.0",
    "PyYAML>=6.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
mcp = [
    "mcp>=1.0.0",
]
kipy = [
    "kicad-python>=0.4.0",  # Official KiCad IPC API (KiCad 9+)
]
streaming = [
    "websockets>=12.0",
]
all = [
    "mcp>=1.0.0",
    "kicad-python>=0.4.0",
    "websockets>=12.0",
]

[project.scripts]
atoplace = "atoplace.cli:main"
atoplace-mcp = "atoplace.mcp.server:main"
atoplace-mcp-launcher = "atoplace.mcp.launcher:main"
atoplace-mcp-bridge = "atoplace.mcp.bridge:main"

[project.urls]
Homepage = "https://github.com/atoplace/atoplace"
Documentation = "https://github.com/atoplace/atoplace#readme"
Repository = "https://github.com/atoplace/atoplace.git"
Issues = "https://github.com/atoplace/atoplace/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["atoplace*"]

[tool.setuptools.package-data]
atoplace = ["*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
